+++ scripts/gauge/components/speedGauge.js
@@ -88,7 +88,7 @@ export function speedGauge(canvasId, tabGrad,unit,gradMin,gradMax,affPosVert,aff
             },
             animation: {
                 delay: 10,
-                duration: 1500,
+                duration: GAUGE_ANIMATION_DURATION,
                 fn: 'linear'
             },
             updateValueOnAnimation: true