+++ scripts/gauge/components/tempEGTGaugeL.js
@@ -85,7 +85,7 @@ export function tempEGTGaugeL(canvasId, tabGrad, unit, gradMin, gradMax, affPosV
 
         animation: {
             delay: 25,
-            duration: 1000,
+            duration: GAUGE_ANIMATION_DURATION,
             //fn : 'bounce'// rebon de l'eguille
             fn: 'linear'
         },