+++ scripts/gauge/components/tempEGTGaugeR.js
@@ -92,7 +92,7 @@ export function tempEGTGaugeR(canvasId, tabGrad, unit, gradMin, gradMax, affPosV
 
         animation: {
             delay: 25,
-            duration: 1000,
+            duration: GAUGE_ANIMATION_DURATION,
             //fn : 'bounce'// rebon de l'eguille
             fn: 'linear'
         },