+++ scripts/gauge/components/tempWaterGaugeL.js
@@ -80,7 +80,7 @@ export function tempWaterGaugeL(canvasId, tabGrad, unit, gradMin, gradMax, affPo
         },
         animation: {
             delay: 25,
-            duration: 1000,
+            duration: GAUGE_ANIMATION_DURATION,
             //fn : 'bounce'// rebon de l'eguille
             fn: 'linear'
         },