+++ scripts/gauge/components/tempWaterGaugeR.js
@@ -88,7 +88,7 @@ export function tempWaterGaugeR(canvasId, tabGrad, unit, gradMin, gradMax, affPo
         },
         animation: {
             delay: 25,
-            duration: 1000,
+            duration: GAUGE_ANIMATION_DURATION,
             //fn : 'bounce'// rebon de l'eguille
             fn: 'linear'
         },